+++ src/main/java/kitchenpos/persistence/MenuGroupRepository.java
@@ -1,5 +1,6 @@
-package kitchenpos.domain;
+package kitchenpos.persistence;
 
+import kitchenpos.domain.MenuGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MenuGroupRepository extends JpaRepository<MenuGroup, Long> {
