+++ src/main/java/kitchenpos/persistence/MenuRepository.java
@@ -1,5 +1,6 @@
-package kitchenpos.domain;
+package kitchenpos.persistence;
 
+import kitchenpos.domain.Menu;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MenuRepository extends JpaRepository<Menu, Long> {
