+++ src/main/java/kitchenpos/persistence/OrderRepository.java
@@ -1,5 +1,7 @@
-package kitchenpos.domain;
+package kitchenpos.persistence;
 
+import kitchenpos.domain.Order;
+import kitchenpos.domain.OrderTable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
