+++ src/main/java/kitchenpos/persistence/OrderTableRepository.java
@@ -1,5 +1,6 @@
-package kitchenpos.domain;
+package kitchenpos.persistence;
 
+import kitchenpos.domain.OrderTable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface OrderTableRepository extends JpaRepository<OrderTable, Long> {
