+++ src/main/java/kitchenpos/persistence/ProductRepository.java
@@ -1,5 +1,6 @@
-package kitchenpos.domain;
+package kitchenpos.persistence;
 
+import kitchenpos.domain.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProductRepository extends JpaRepository<Product, Long> {
