+++ src/main/java/kitchenpos/persistence/TableGroupRepository.java
@@ -1,5 +1,6 @@
-package kitchenpos.domain;
+package kitchenpos.persistence;
 
+import kitchenpos.domain.TableGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TableGroupRepository extends JpaRepository<TableGroup, Long> {
